# telehealth-bangladesh/API's/telehealth_openapi_spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HealNSight Telehealth Bangladesh API",
    "description": "Complete OpenAPI 3.0.3 specification covering all 58 endpoints across Django REST Framework, Node.js Auth Microservice, and FastAPI Worker.",
    "version": "1.0.0",
    "contact": {
      "name": "HealNSight Technical Team",
      "url": "https://github.com/infectedsourav/telehealth-bangladesh"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Django REST Backend (Core Clinical & Operations)"
    },
    {
      "url": "http://localhost:5000",
      "description": "Node.js Auth & OTP Microservice"
    },
    {
      "url": "http://localhost:6000",
      "description": "FastAPI AI & Vision Worker"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token (obtainable from /api/login/ or /api/auth/login)."
      }
    }
  },
  "paths": {
...
    "/api/telephony/call/": {
      "post": {
        "tags": [
          "Telephony"
        ],
        "summary": "Initiate Cellular Phone Bridge",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Validation or format error"
          },
          "401": {
            "description": "Authentication credentials missing or invalid"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example": {
                  "doctor_phone": "+8801700000002",
                  "patient_phone": "+8801700000001",
                  "consultation_id": 1
                }
              }
            }
          }
        }
      }
    },
    "/api/telephony/call/{call_sid}/terminate/": {
      "post": {
        "tags": [
          "Telephony"
        ],
        "summary": "Terminate Active Voice Call",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Validation or format error"
          },
          "401": {
            "description": "Authentication credentials missing or invalid"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/telephony/call/{call_sid}/status/": {
      "get": {
        "tags": [
          "Telephony"
        ],
        "summary": "Get Call Status",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Validation or format error"
          },
          "401": {
            "description": "Authentication credentials missing or invalid"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/telephony/twiml/": {
      "post": {
        "tags": [
          "Telephony"
        ],
        "summary": "Dynamic Voice TwiML Instructions",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Validation or format error"
          },
          "401": {
            "description": "Authentication credentials missing or invalid"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/telephony/twiml/{consultation_id}/": {
      "post": {
        "tags": [
          "Telephony"
        ],
        "summary": "Consultation Conference TwiML",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Validation or format error"
          },
          "401": {
            "description": "Authentication credentials missing or invalid"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  }
}